refactor(perl-lexer): split core lexer into srp module boundary (#0000)
